requirements.txt
aniso8601==1.1.0
backports.ssl-match-hostname==3.5.0.1
boto3==1.3.1
botocore==1.4.25
cached-property==1.3.0
click==6.6
docker-compose==1.7.1
docker-py==1.8.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
enum34==1.1.6
Flask==0.11.1
Flask-RESTful==0.3.5
functools32==3.2.3.post2
futures==3.0.5
gevent==1.1.1
greenlet==0.4.10
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
jsonschema==2.5.1
MarkupSafe==0.23
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
requests==2.7.0
six==1.10.0
texttable==0.8.4
websocket-client==0.37.0
Werkzeug==0.11.10
wheel==0.24.0
